+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+----------
+### Added
+- support for changing the brightness of all connected displays simultaneously
+
 [1.0.0](../../releases/tag/v1.0.0) - 2019-08-28
 -----------------------------------------------
 Initial release of the app.